mech graphics pt1
everything but the mech!!!!!
the beam flash should spawn and loop at the tip of the rifle just like
how you have the pilot's effect going.
gatling bullet spawns near the head, it fires out of the ears/hood.
do whatever you want with the beam, if you had a better idea for it then
let me know